\name{mrloglin}
\alias{mrloglin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fit loglinear models to multiple-response data
}
\description{
Fits loglinear models to the multiway contingency tables generated by multiple-response data, treating them as clustered factor data. Standard errors use a sandwich estimator; the Rao-Scott-type score and deviance tests are used to compare models.
}
\usage{
mrloglin(formula, data, ...)
update(model, formula, ...)
anova(model1, model2,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{formula}{
  one-sided model formula for a loglinear model, or formula update for a new model
}
  \item{data}{
  data frame containing all the variables
}
\item{model}{model fitted by \code{mrloglin}}
\item{model1,model2}{Two \code{mrloglin} models to compare. These must have the same variables in the same order.}
  \item{\dots}{
not used
}
}
\details{
The tests are the same as those of Decady and Thomas, which are the same as those of Loughin and Scherer except that they use the asymptotic reference distribution rather than a bootstrap. 
}
\value{
Object of type \code{mrloglin}
}
\references{

Decady YJ, Thomas DR (2000) "A Simple Test of Association for Contingency Tables with Multiple Column Responses" Biometrics 56: 893-896

Loughin TM, Scherer PN (1998) "Testing for Association in Contingency Tables with Multiple Column Responses" Biometrics 54: 630-637

Rao, JNK, Scott, AJ (1984) "On Chi-squared Tests For Multiway Contigency Tables with Proportions Estimated From Survey Data" Annals of Statistics 12:46-60.}

\note{
The underlying code in \code{svyloglin} is a bit fragile. Real examples of it failing would be useful. 
}


\seealso{
\code{\link{mrglm}} for an alternative modelling approach

The vignette "Working models for multiple response data"

\code{\link[survey]{svyloglin}}
}
\examples{
data(vetinfo)

m0<-mrloglin(~education+sources, data=vetinfo)
m1<-mrloglin(~education*sources, data=vetinfo)
summary(m0)
summary(m1)
anova(m0,m1)


}