\name{mrloglin}
\alias{mrloglin}
\alias{update.mrloglin}
\alias{print.anova.mrloglin}
\alias{anova.mrloglin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fit loglinear models to multiple-response data
}
\description{
Fits loglinear models to the multiway contingency tables generated by multiple-response data, treating them as clustered factor data. Standard errors use a sandwich estimator; the Rao-Scott-type score and deviance tests are used to compare models.
}
\usage{
mrloglin(formula, data, ...)
\method{update}{mrloglin}(model, formula, ...)
\method{anova}{mrloglin}(model1, model2,...,integrate=FALSE)
\method{print}{anova.mrloglin}(x,pval=c("F","saddlepoint","lincom","chisq"),...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{formula}{
  one-sided model formula for a loglinear model, or formula update for a new model
}
  \item{data}{
  data frame containing all the variables
}
\item{model}{model fitted by \code{mrloglin}}
\item{model1,model2}{Two \code{mrloglin} models to compare. These must have the same variables in the same order.}
\item{integrate}{Compute the exact linear combination of chisquared distributions that gives the estimated asymptotic distribution}
  \item{\dots}{
not used
}
}
\details{
Loughin and Scherer used the same test statistic but with a bootstrap null distribution. Decady and Thomas noted that this setting was a special case of the Rao-Scott tests in survey sampling and used the first-order Rao-Scott correction, scaling the test statistic to have the correct mean.  The \code{print} method has an option for the p-value approximation to be used:   the exact linear combination of chisquared (only if \code{integrate=TRUE}), the first and second order Rao-Scott corrections (\code{"chisq"} and \code{"F"}), and the saddlepoint approximation (\code{"saddlepoint"}). The default is the second-order Rao-Scott correction.

}
\value{
Object of type \code{mrloglin}
}
\references{

Decady YJ, Thomas DR (2000) "A Simple Test of Association for Contingency Tables with Multiple Column Responses" Biometrics 56: 893-896

Loughin TM, Scherer PN (1998) "Testing for Association in Contingency Tables with Multiple Column Responses" Biometrics 54: 630-637

Rao, JNK, Scott, AJ (1984) "On Chi-squared Tests For Multiway Contigency Tables with Proportions Estimated From Survey Data" Annals of Statistics 12:46-60.}

\note{
The underlying code in \code{svyloglin} is a bit fragile. Real examples of it failing would be useful. 
}


\seealso{
\code{\link{mrglm}} for an alternative modelling approach, \code{\link{mrmultinom}} for multinomial regression models

The vignette "Working models for multiple response data"

\code{\link{svyloglin}}
}
\examples{
data(vetinfo)

m0<-mrloglin(~education+sources, data=vetinfo)
m1<-mrloglin(~education*sources, data=vetinfo)
summary(m0)
summary(m1)
anova(m0,m1)


}